Move MoE Router related code to separate files
